Update robot container for all subsystems and commands

+++ src/main/java/frc/robot/commands/ElevatorControlCommand.java
@@ -6,9 +6,9 @@ import edu.wpi.first.wpilibj2.command.Command;
 
 public class ElevatorControlCommand extends Command {
     private final ElevatorSubsystem elevator;
-    private final int level;
+    private final double level;
 
-    public ElevatorControlCommand(ElevatorSubsystem elevator, int level) {
+    public ElevatorControlCommand(ElevatorSubsystem elevator, double level) {
         this.elevator = elevator;
         this.level = level;
         addRequirements(elevator);
